+++ chain/src/indexed_block.rs
@@ -86,7 +86,7 @@ impl IndexedBlock {
 		let hashes = match self.transactions.split_first() {
 			None => vec![],
 			Some((_, rest)) => {
-				let mut hashes = vec![H256::from(0)];
+				let mut hashes = vec![H256::from(0 as u8)];
 				hashes.extend(rest.iter().map(|tx| tx.raw.witness_hash()));
 				hashes
 			},
